Split Constants per team
